+++ libmtx/mtxfile/data.h
@@ -106,7 +106,7 @@ int mtxfile_data_size_per_element(
  */
 int mtxfile_parse_data_array_real_single(
     float * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -123,7 +123,7 @@ int mtxfile_parse_data_array_real_single(
  */
 int mtxfile_parse_data_array_real_double(
     double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -140,7 +140,7 @@ int mtxfile_parse_data_array_real_double(
  */
 int mtxfile_parse_data_array_complex_single(
     float (* data)[2],
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -157,7 +157,7 @@ int mtxfile_parse_data_array_complex_single(
  */
 int mtxfile_parse_data_array_complex_double(
     double (* data)[2],
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -174,7 +174,7 @@ int mtxfile_parse_data_array_complex_double(
  */
 int mtxfile_parse_data_array_integer_single(
     int32_t * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -191,7 +191,7 @@ int mtxfile_parse_data_array_integer_single(
  */
 int mtxfile_parse_data_array_integer_double(
     int64_t * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -212,7 +212,7 @@ int mtxfile_parse_data_array_integer_double(
  */
 int mtxfile_parse_data_matrix_coordinate_real_single(
     struct mtxfile_matrix_coordinate_real_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -231,7 +231,7 @@ int mtxfile_parse_data_matrix_coordinate_real_single(
  */
 int mtxfile_parse_data_matrix_coordinate_real_double(
     struct mtxfile_matrix_coordinate_real_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -250,7 +250,7 @@ int mtxfile_parse_data_matrix_coordinate_real_double(
  */
 int mtxfile_parse_data_matrix_coordinate_complex_single(
     struct mtxfile_matrix_coordinate_complex_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -269,7 +269,7 @@ int mtxfile_parse_data_matrix_coordinate_complex_single(
  */
 int mtxfile_parse_data_matrix_coordinate_complex_double(
     struct mtxfile_matrix_coordinate_complex_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -288,7 +288,7 @@ int mtxfile_parse_data_matrix_coordinate_complex_double(
  */
 int mtxfile_parse_data_matrix_coordinate_integer_single(
     struct mtxfile_matrix_coordinate_integer_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -307,7 +307,7 @@ int mtxfile_parse_data_matrix_coordinate_integer_single(
  */
 int mtxfile_parse_data_matrix_coordinate_integer_double(
     struct mtxfile_matrix_coordinate_integer_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -326,7 +326,7 @@ int mtxfile_parse_data_matrix_coordinate_integer_double(
  */
 int mtxfile_parse_data_matrix_coordinate_pattern(
     struct mtxfile_matrix_coordinate_pattern * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows,
@@ -349,7 +349,7 @@ int mtxfile_parse_data_matrix_coordinate_pattern(
  */
 int mtxfile_parse_data_vector_coordinate_real_single(
     struct mtxfile_vector_coordinate_real_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -367,7 +367,7 @@ int mtxfile_parse_data_vector_coordinate_real_single(
  */
 int mtxfile_parse_data_vector_coordinate_real_double(
     struct mtxfile_vector_coordinate_real_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -385,7 +385,7 @@ int mtxfile_parse_data_vector_coordinate_real_double(
  */
 int mtxfile_parse_data_vector_coordinate_complex_single(
     struct mtxfile_vector_coordinate_complex_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -403,7 +403,7 @@ int mtxfile_parse_data_vector_coordinate_complex_single(
  */
 int mtxfile_parse_data_vector_coordinate_complex_double(
     struct mtxfile_vector_coordinate_complex_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -421,7 +421,7 @@ int mtxfile_parse_data_vector_coordinate_complex_double(
  */
 int mtxfile_parse_data_vector_coordinate_integer_single(
     struct mtxfile_vector_coordinate_integer_single * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -439,7 +439,7 @@ int mtxfile_parse_data_vector_coordinate_integer_single(
  */
 int mtxfile_parse_data_vector_coordinate_integer_double(
     struct mtxfile_vector_coordinate_integer_double * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -457,7 +457,7 @@ int mtxfile_parse_data_vector_coordinate_integer_double(
  */
 int mtxfile_parse_data_vector_coordinate_pattern(
     struct mtxfile_vector_coordinate_pattern * data,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     int num_rows);
@@ -525,7 +525,7 @@ int mtxfile_fread_data(
     union mtxfile_data * data,
     FILE * f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
     enum mtxfile_object object,
@@ -556,7 +556,7 @@ int mtxfile_gzread_data(
     union mtxfile_data * data,
     gzFile f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf,
     enum mtxfile_object object,
