+++ libmtx/mtxfile/header.h
@@ -121,7 +121,7 @@ const char * mtxfile_object_str(
  */
 int mtxfile_parse_object(
     enum mtxfile_object * object,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     const char * valid_delimiters);
@@ -156,7 +156,7 @@ const char * mtxfile_format_str(
  */
 int mtxfile_parse_format(
     enum mtxfile_format * format,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     const char * valid_delimiters);
@@ -191,7 +191,7 @@ const char * mtxfile_field_str(
  */
 int mtxfile_parse_field(
     enum mtxfile_field * field,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     const char * valid_delimiters);
@@ -226,7 +226,7 @@ const char * mtxfile_symmetry_str(
  */
 int mtxfile_parse_symmetry(
     enum mtxfile_symmetry * symmetry,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s,
     const char * valid_delimiters);
@@ -285,7 +285,7 @@ struct mtxfile_header
  */
 int mtxfile_parse_header(
     struct mtxfile_header * header,
-    int * bytes_read,
+    int64_t * bytes_read,
     const char ** endptr,
     const char * s);
 
@@ -312,7 +312,7 @@ int mtxfile_fread_header(
     struct mtxfile_header * header,
     FILE * f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
 
@@ -329,7 +329,7 @@ int mtxfile_gzread_header(
     struct mtxfile_header * header,
     gzFile f,
     int * lines_read,
-    int * bytes_read,
+    int64_t * bytes_read,
     size_t line_max,
     char * linebuf);
 #endif
